Refactor: move non-sms messagebatch classes to new messagebatch package
