+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/program/message/ProgramMessageService.java
@@ -32,7 +32,7 @@ import java.util.Date;
 import java.util.List;
 import java.util.Set;
 
-import org.hisp.dhis.sms.BatchResponseStatus;
+import org.hisp.dhis.messagebatch.BatchResponseStatus;
 import org.hisp.dhis.user.User;
 
 /**
