+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/sms/SmsMessageBatchCreatorService.java
@@ -32,8 +32,10 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.hisp.dhis.common.DeliveryChannel;
+import org.hisp.dhis.messagebatch.MessageBatchCreatorService;
+import org.hisp.dhis.messagebatch.OutBoundMessage;
 import org.hisp.dhis.program.message.ProgramMessage;
-import org.hisp.dhis.sms.outbound.MessageBatch;
+import org.hisp.dhis.messagebatch.OutboundMessageBatch;
 
 /**
  * @author Zubair <rajazubair.asghar@gmail.com>
@@ -42,9 +44,9 @@ public class SmsMessageBatchCreatorService
     implements MessageBatchCreatorService
 {
     @Override
-    public MessageBatch getMessageBatch( List<ProgramMessage> programMessages )
+    public OutboundMessageBatch getMessageBatch( List<ProgramMessage> programMessages )
     {
-        MessageBatch messageBatch = new MessageBatch();
+        OutboundMessageBatch messageBatch = new OutboundMessageBatch();
 
         List<OutBoundMessage> smsBatch = new ArrayList<>();
